# src/swagger/swagger.yml
openapi: 3.0.0
paths:
...
  /menu:
    get:
      operationId: Menucontroller_getAllMenuById
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_1
        - Menu
    post:
      operationId: Menucontroller_createMenu
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
  /menu/{id}/menuitems:
    get:
      operationId: Menucontroller_getMenuItemsById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_1
  /menu/{id}/menuitem:
    post:
      operationId: Menucontroller_createMenuItem
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemDto'
      responses:
        '201':
          description: ''
      tags: *ref_1
...
  /Orders/order:
    post:
      operationId: OrderController_createOrder
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createOrderDTo'
      responses:
        '201':
          description: ''
      tags: &ref_5
        - Orders
  /Orders/order/{id}:
    get:
      operationId: OrderController_getOrderDetailsById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_5
  /Orders/ordername/{name}:
    get:
      operationId: OrderController_getOrderDetailsByName
      parameters:
        - name: name
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_5
  /Orders/updatordername/{name}:
    put:
      operationId: OrderController_updateOrderQuantity
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateOrderDto'
      responses:
        '200':
          description: ''
      tags: *ref_5
  /Orders/menuItem:
    delete:
      operationId: OrderController_deleteMenuItem
      parameters:
        - name: OrderId
          required: true
          in: query
          schema:
            type: number
        - name: menuItem
          required: true
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_5
...
  /auth/employee/login:
    post:
      operationId: AuthController_employeeLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginEmployeeDto'
      responses:
        '201':
          description: ''
      tags: &ref_9
        - employee
  /auth/employee/register:
    post:
      operationId: AuthController_employeeRegister
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerEmployeeDto'
      responses:
        '201':
          description: ''
      tags: *ref_9
  /auth/employee/{id}/details:
    get:
      operationId: AuthController_getEmployeeDetails
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_9
  /auth/employee/roles:
    post:
      operationId: AuthController_createRoles
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createRoleDto'
      responses:
        '201':
          description: ''
      tags: *ref_9
  /auth/employee/updaterole/{id}:
    put:
      operationId: AuthController_updateRoles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateRoleDto'
      responses:
        '200':
          description: ''
      tags: *ref_9
info:
  title: chandu
  description: Your API Description
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  schemas:
    MenuDto:
      type: object
      properties: {}
    MenuItemDto:
      type: object
      properties:
        menu_itemname:
          type: string
        price:
          type: number
      required:
        - menu_itemname
        - price
...
    createOrderDTo:
      type: object
      properties:
        customerName:
          type: string
        items:
          type: array
          items:
            type: string
      required:
        - customerName
        - items
    updateOrderDto:
      type: object
      properties:
        menuItem:
          type: string
        quantity:
          type: number
      required:
        - menuItem
        - quantity
...
    loginEmployeeDto:
      type: object
      properties: {}
    registerEmployeeDto:
      type: object
      properties: {}
    createRoleDto:
      type: object
      properties: {}
    updateRoleDto:
      type: object
      properties: {}